+++ src/index.js
@@ -22,7 +22,6 @@ exports.directory = {
  */
 
 exports.forge = require('./crypto/forge');
-exports.openssl = require('./crypto/openssl');
 
 
 /**